+++ src/main/java/com/beyond/pochaon/present/dto/PresentCreateDto.java
@@ -14,8 +14,8 @@ import java.util.UUID;
 public class PresentCreateDto {
     private UUID idempotencyKey;
 
-    private Long senderTableNum;
-    private Long receiverTableNum;
+    private int senderTableNum;
+    private int receiverTableNum;
     private Long menuId;
     private int menuQuantity;
 }
