+++ src/main/java/com/beyond/pochaon/present/dto/PresentOwnerDto.java
@@ -15,9 +15,9 @@ import java.util.UUID;
 @NoArgsConstructor
 @Data
 public class PresentOwnerDto {
-    private Long senderTableNum;
+    private int senderTableNum;
     private UUID groupId;
-    private Long receiverTableNum;
+    private int receiverTableNum;
     private List<MenuDto> menuDtoList;
 
     @Builder
